+++ packages/everything-web/packages/react-ui/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2-alpha
+
+-   Pagination component
+-   VideoList component
+
 ## 0.0.1-alpha
 
 -   fix: colors used as `grey` instead of `gray`